Remove recommendation in favour of better style
I followed the link from the Save Actions page, and found this:
https://tex.stackexchange.com/a/10775
According to that post, protecting just the first letter of a word, might lead to bad spacing after the capital letter.
This might be a bit perfectionist, but I would suggest leaving the recommendation in brackets out, to be consistent.