Add Linux Mint LibreOffice DEB Application Path
When LibreOffice was manually installed via a deb file, the path is in an odd location and hard to find. I added this path to the documentation. Usually, it is not recommended for users of Linux Mint to install deb files manually (The package manager and Flatpak are better choices), but some users on long term supported distributions may be tempted to install the newest version of LibreOffice that way, as the package manager may not feature new versions.